Move the Result family out of command.
This code is now more widely used than in just leaf command
implementations.
