+++ pex/resolve/lockfile/updater.py
@@ -8,7 +8,6 @@ import os
 from collections import OrderedDict
 from contextlib import contextmanager
 
-from pex.commands.command import Error, ResultError, catch, try_
 from pex.common import pluralize
 from pex.network_configuration import NetworkConfiguration
 from pex.pep_440 import Version
@@ -17,6 +16,7 @@ from pex.resolve.locked_resolve import LockConfiguration, LockedRequirement, Loc
 from pex.resolve.lockfile import Lockfile, create
 from pex.resolve.requirement_configuration import RequirementConfiguration
 from pex.resolve.resolver_configuration import PipConfiguration, ReposConfiguration
+from pex.result import Error, ResultError, catch, try_
 from pex.sorted_tuple import SortedTuple
 from pex.targets import AbbreviatedPlatform, LocalInterpreter, Target, Targets
 from pex.third_party.packaging import tags
