+++ pex/tools/commands/graph.py
@@ -10,10 +10,11 @@ import threading
 from argparse import ArgumentParser
 from contextlib import contextmanager
 
-from pex.commands.command import Ok, OutputMixin, Result, try_open_file, try_run_program
+from pex.commands.command import OutputMixin, try_open_file, try_run_program
 from pex.common import safe_mkdir
 from pex.dist_metadata import requires_dists
 from pex.pex import PEX
+from pex.result import Ok, Result
 from pex.tools.command import PEXCommand
 from pex.tools.commands.digraph import DiGraph
 from pex.typing import TYPE_CHECKING
