+++ pex/tools/commands/repository.py
@@ -15,7 +15,7 @@ from textwrap import dedent
 from threading import Thread
 
 from pex import dist_metadata
-from pex.commands.command import Error, JsonMixin, Ok, OutputMixin, Result
+from pex.commands.command import JsonMixin, OutputMixin
 from pex.common import (
     DETERMINISTIC_DATETIME_TIMESTAMP,
     pluralize,
@@ -28,6 +28,7 @@ from pex.environment import PEXEnvironment
 from pex.interpreter import PythonIdentity, PythonInterpreter, spawn_python_job
 from pex.jobs import Retain, SpawnedJob, execute_parallel
 from pex.pex import PEX
+from pex.result import Error, Ok, Result
 from pex.third_party.pkg_resources import Distribution
 from pex.tools.command import PEXCommand
 from pex.typing import TYPE_CHECKING, cast
