+++ pex/tools/commands/venv.py
@@ -13,13 +13,13 @@ from collections import Counter, defaultdict
 from textwrap import dedent
 
 from pex import pex_warnings
-from pex.commands.command import Error, Ok, Result
 from pex.common import chmod_plus_x, pluralize, safe_delete, safe_mkdir, safe_rmtree
 from pex.compatibility import is_valid_python_identifier
 from pex.enum import Enum
 from pex.environment import PEXEnvironment
 from pex.orderedset import OrderedSet
 from pex.pex import PEX
+from pex.result import Error, Ok, Result
 from pex.tools.command import PEXCommand
 from pex.tools.commands.virtualenv import PipUnavailableError, Virtualenv
 from pex.tracer import TRACER
